# lib/api-zod/src/generated/types/createRecordingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateRecordingBodyBrowserInfo } from "./createRecordingBodyBrowserInfo";
import type { CreateRecordingBodyThumbnail } from "./createRecordingBodyThumbnail";
import type { NetworkLogEntry } from "./networkLogEntry";

export interface CreateRecordingBody {
  title: string;
  duration: number;
  pageUrl?: string | null;
  pageTitle?: string | null;
  tags?: string[];
  events: NetworkLogEntry[];
  videoObjectPath?: string | null;
  thumbnailObjectPath?: string | null;
  thumbnail?: CreateRecordingBodyThumbnail;
  browserInfo?: CreateRecordingBodyBrowserInfo;
}