# lib/api-zod/src/generated/types/recording.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RecordingBrowserInfo } from "./recordingBrowserInfo";

export interface Recording {
  id: string;
  userId: string;
  title: string;
  /** Duration in milliseconds */
  duration: number;
  createdAt: Date;
  pageUrl?: string | null;
  pageTitle?: string | null;
  networkLogsCount: number;
  errorCount: number;
  consoleCount: number;
  clickCount: number;
  videoObjectPath?: string | null;
  thumbnailObjectPath?: string | null;
  shareToken?: string | null;
  tags: string[];
  browserInfo?: RecordingBrowserInfo;
}